pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clinical-rag-extractor"
version = "0.1.0"
description = "Hallucination-mitigated clinical data extraction using RAG with structured output enforcement"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "DanielMartin Arogyasami"}]
keywords = ["clinical-nlp", "rag", "hallucination-mitigation", "sdtm", "meddra", "loinc"]

dependencies = [
    "litellm>=1.40.0",
    "openai>=1.30.0",
    "instructor>=1.3.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "transformers>=4.40.0",
    "torch>=2.2.0",
    "pydantic>=2.7.0",
    "jsonschema>=4.22.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scipy>=1.13.0",
    "scikit-learn>=1.5.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    "httpx>=0.27.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests requiring ML model downloads (deselect with '-m \"not slow\"')",
]
